fix: refactor marks form to seperate component
