+++ app/(tabs)/siktemerker/components/MarksTable.tsx
@@ -1,8 +1,8 @@
 import { faTrash } from '@fortawesome/free-solid-svg-icons/faTrash';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { StyleSheet, Text, View } from 'react-native';
-import Button from '../../../components/common/Button';
-import { CalculatedMarks } from '../../../types';
+import Button from '../../../../components/common/Button';
+import { CalculatedMarks } from '../../../../types';
 
 interface CalculationTableProps {
   ballistics: CalculatedMarks | null;
