+++ app/(tabs)/siktemerker/index.tsx
@@ -2,16 +2,14 @@ import * as Sentry from '@sentry/react-native';
 import { useEffect, useState } from 'react';
 import { Keyboard, StyleSheet, Text, TouchableWithoutFeedback, View } from 'react-native';
 
-import { Button, Input } from '../../../components/common';
 import { AimDistanceMark, CalculatedMarks, MarkValue } from '../../../types';
-import { Ballistics, formatNumber, getLocalStorage, storeLocalStorage, useBallisticsParams } from '../../../utils/';
-import MarksTable from './MarksTable';
-import { useCalcForm } from './useCalcForm';
+import { Ballistics, getLocalStorage, storeLocalStorage, useBallisticsParams } from '../../../utils/';
+import MarksForm from './components/MarksForm';
+import MarksTable from './components/MarksTable';
 
 export default function Calculate() {
   const [ballistics, setBallistics] = useState<CalculatedMarks | null>(null);
   const { error, status, calculateBallisticsParams } = useBallisticsParams();
-  const [{ aimError, aimValue, distanceError, distanceValue }, dispatch] = useCalcForm();
 
   useEffect(() => {
     getLocalStorage<CalculatedMarks>('ballistics').then((data) => {
@@ -46,31 +44,6 @@ export default function Calculate() {
     }
   }
 
-  function handleDistanceChange(value: string) {
-    dispatch({ type: 'SET_DISTANCE_VALUE', payload: value });
-  }
-
-  function handleAimChange(value: string) {
-    dispatch({ type: 'SET_AIM_VALUE', payload: value });
-  }
-
-  async function handleAddMark() {
-    if (!aimValue) {
-      dispatch({ type: 'SET_AIM_ERROR', payload: true });
-    }
-    if (!distanceValue) {
-      dispatch({ type: 'SET_DISTANCE_ERROR', payload: true });
-    }
-    if (aimValue && distanceValue) {
-      const newEntry: MarkValue = { aim: parseFloat(aimValue), distance: parseFloat(distanceValue) };
-
-      await sendMarks(newEntry);
-      dispatch({ type: 'SET_AIM_VALUE', payload: '' });
-      dispatch({ type: 'SET_DISTANCE_VALUE', payload: '' });
-      Keyboard.dismiss();
-    }
-  }
-
   async function handleRemoveMark(index: number) {
     const newDistances = ballistics.given_distances.filter((_distance: any, i: number) => i === index);
 
@@ -81,44 +54,7 @@ export default function Calculate() {
     <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
       <View style={{ flex: 1 }}>
         <Text style={styles.title}>Siktemerker</Text>
-        <View style={styles.form}>
-          <View>
-            <Input
-              textAlign="right"
-              maxLength={100}
-              label="Avstand"
-              onBlur={() => dispatch({ type: 'SET_DISTANCE_ERROR', payload: false })}
-              placeholderText="F.eks. 20"
-              keyboardType="numeric"
-              error={distanceError}
-              errorMessage="Fyll inn avstand"
-              value={distanceValue}
-              onChangeText={(value) => handleDistanceChange(formatNumber(value))}
-            />
-          </View>
-          <View>
-            <Input
-              textAlign="right"
-              maxLength={15}
-              label="Merke"
-              onBlur={() => dispatch({ type: 'SET_AIM_ERROR', payload: false })}
-              placeholderText="F.eks. 2.35"
-              keyboardType="numeric"
-              value={aimValue}
-              error={aimError}
-              errorMessage="Fyll inn siktemerke"
-              onChangeText={(value) => handleAimChange(formatNumber(value))}
-            />
-          </View>
-          <Button
-            type="filled"
-            width={100}
-            loading={status === 'pending'}
-            buttonStyle={{ marginLeft: 'auto', marginTop: 16 }}
-            onPress={handleAddMark}
-            label="Beregn"
-          />
-        </View>
+        <MarksForm sendMarks={sendMarks} status={status} />
         {error && (
           <>
             <View style={{ marginBottom: 8, padding: 8 }}>Obs, noe gikk galt. Prøv igjen senere.</View>
@@ -135,12 +71,5 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     marginBottom: 16,
     marginTop: 16,
-  },
-  form: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 16,
   },
 });
